Portuguese application xml - last pt translation file
This is the translation of the application XML file from English
to Portuguese.

FBReader already had some Portuguese language assets such as hypenation
patterns and language patterns.  With this third of three commits,
FBReader should have everything it needs in terms of translation for
Portuguese speaking users, as far as know any how.

Thanks to http://www.gototranslation.com for the translation to
Portuguese.